chore(footer): relocated component
