+++ next-tavla/app/(admin)/organizations/[id]/page.tsx
@@ -11,7 +11,7 @@ import { UploadLogo } from 'Admin/scenarios/Organization/components/UploadLogo'
 import { MemberAdministration } from 'Admin/scenarios/Organization/components/MemberAdministration'
 import { permanentRedirect } from 'next/navigation'
 import { getUserFromSessionCookie } from 'Admin/utils/formActions'
-import { Info } from 'Admin/scenarios/Organization/components/Info'
+import { Info } from './components/Info'
 
 initializeAdminApp()
