--- next-tavla/src/Admin/scenarios/Organization/components/Info/styles.module.css
@@ -1,5 +1,0 @@
-.infoForm {
-    display: flex;
-    gap: 0.5em;
-    height: 4rem;
-}
